refactor: add react-refresh eslint rule; fix one instance
