+++ src/contexts/ThemeProvider/customRender.ts
@@ -1,0 +1,9 @@
+import type { RenderOptions } from '@testing-library/react'
+import { render } from '@testing-library/react'
+import type { ReactElement } from 'react'
+import { AllTheProviders } from './test-utils'
+
+const customRender = (ui: ReactElement, options?: Omit<RenderOptions, 'wrapper'>) =>
+  render(ui, { wrapper: AllTheProviders, ...options })
+
+export { customRender as render }
