+++ src/contexts/ThemeProvider/index.ts
@@ -1,0 +1,3 @@
+export { ThemeProvider, type ThemeProviderProps } from './ThemeProvider'
+export { useTheme } from './useTheme'
+export { render } from './customRender'
