+++ src/contexts/ThemeProvider/test-utils.tsx
@@ -1,13 +1,10 @@
-import type { RenderOptions } from '@testing-library/react'
-import { render } from '@testing-library/react'
-import type { ReactElement } from 'react'
 import type React from 'react'
 import { I18nextProvider, useTranslation } from 'react-i18next'
 import { createTheme } from './createTheme'
 import { ThemeProvider } from './ThemeProvider'
 import '@/i18n'
 
-const AllTheProviders = ({ children }: { children: React.ReactNode }) => {
+export const AllTheProviders = ({ children }: { children: React.ReactNode }) => {
   const { i18n } = useTranslation()
   const theme = createTheme()
   return (
@@ -16,8 +13,3 @@ const AllTheProviders = ({ children }: { children: React.ReactNode }) => {
     </ThemeProvider>
   )
 }
-
-const customRender = (ui: ReactElement, options?: Omit<RenderOptions, 'wrapper'>) =>
-  render(ui, { wrapper: AllTheProviders, ...options })
-
-export { customRender as render }
