+++ src/contexts/ThemeProvider/useTheme.ts
@@ -1,0 +1,8 @@
+import { createContext, useContext } from 'react'
+
+export interface ThemeContextProps {
+  container: React.RefObject<HTMLElement>
+}
+export const ThemeContext = createContext<ThemeContextProps>({} as ThemeContextProps)
+
+export const useTheme = () => useContext(ThemeContext)
